+++ torch/distributed/rpc_api.py
@@ -3,7 +3,7 @@ from . import init_rref_context
 from . import ProcessGroupAgent
 from . import WorkerId
 import torch.distributed.rpc as rpc
-from .internal_rpc_utils import _internal_rpc_pickler, PythonUDF
+from torch.distributed.rpc.internal import _internal_rpc_pickler, PythonUDF
 
 import functools
 import sys
